+++ StripePayments/StripePayments/Source/PaymentHandler/STPPaymentHandler.swift
@@ -644,6 +644,7 @@ public class STPPaymentHandler: NSObject {
             .amazonPay,
             .alma,
             .konbini:
+            .promptPay:
             return false
 
         case .unknown:
@@ -1355,6 +1356,21 @@ public class STPPaymentHandler: NSObject {
                     )
                 )
             }
+        case .promptpayDisplayQrCode:
+            guard
+                let returnURL = URL(string: currentAction.returnURLString ?? ""),
+                let presentingVC = currentAction.authenticationContext
+                    as? PaymentSheetAuthenticationContext,
+                let hostedInstructionsURL = authenticationAction.promptPayDisplayQrCode?.hostedInstructionsURL
+
+            else {
+                fatalError()
+            }
+
+            _handleRedirect(to: hostedInstructionsURL, fallbackURL: hostedInstructionsURL, return: returnURL) { safariViewController in
+                // Present the polling view controller behind the web view so we can start polling right away
+                presentingVC.presentPollingVCForAction(action: currentAction, type: .promptPay, safariViewController: safariViewController)
+            }
         @unknown default:
             fatalError()
         }
@@ -1479,7 +1495,8 @@ public class STPPaymentHandler: NSObject {
                                                     retryCount: retryCount - 1
                                                 )
                                             }
-                                        } else if retrievedPaymentIntent?.paymentMethod?.type != .paynow {
+                                        } else if retrievedPaymentIntent?.paymentMethod?.type != .paynow
+                                                    && retrievedPaymentIntent?.paymentMethod?.type != .promptPay {
                                             // For PayNow, we don't want to mark as canceled when the web view dismisses
                                             // Instead we rely on the presented PollingViewController to complete the currentAction
                                             self._markChallengeCanceled(withCompletion: { _, _ in
@@ -1803,7 +1820,8 @@ public class STPPaymentHandler: NSObject {
                 .alipayHandleRedirect,
                 .weChatPayRedirectToApp,
                 .cashAppRedirectToApp,
-                .payNowDisplayQrCode:
+                .payNowDisplayQrCode,
+                .promptpayDisplayQrCode:
                 return false
             case .OXXODisplayDetails,
                 .boletoDisplayDetails,
@@ -1835,7 +1853,7 @@ public class STPPaymentHandler: NSObject {
             threeDSSourceID = nextAction.useStripeSDK?.threeDSSourceID
         case .OXXODisplayDetails, .alipayHandleRedirect, .unknown, .BLIKAuthorize,
             .weChatPayRedirectToApp, .boletoDisplayDetails, .verifyWithMicrodeposits,
-            .upiAwaitNotification, .cashAppRedirectToApp, .konbiniDisplayDetails, .payNowDisplayQrCode:
+            .upiAwaitNotification, .cashAppRedirectToApp, .konbiniDisplayDetails, .payNowDisplayQrCode, .promptpayDisplayQrCode:
             break
         @unknown default:
             fatalError()